2020/06/11 Jboard1 실습추가

+++ Jboard1/WebContent/view.jsp
@@ -1,3 +1,5 @@
+<%@page import="java.util.ArrayList"%>
+<%@page import="java.util.List"%>
 <%@page import="kr.co.jboard1.bean.ArticleBean"%>
 <%@page import="java.sql.ResultSet"%>
 <%@page import="kr.co.jboard1.config.SQL"%>
@@ -28,9 +30,13 @@
 	PreparedStatement psmt = conn.prepareStatement(SQL.SELECT_ARTICLE);
 	psmt.setString(1, seq);
 	
+	PreparedStatement psmtComment = conn.prepareStatement(SQL.SELECT_COMMENTS);
+	psmtComment.setString(1, seq);
+	
 	// 4단계
 	psmtHit.executeUpdate();
 	ResultSet rs = psmt.executeQuery();
+	ResultSet rsComment = psmtComment.executeQuery();
 	
 	// 5단계
 	ArticleBean article = new ArticleBean();
@@ -49,11 +55,31 @@
 		article.setRdate(rs.getString(11));
 	}
 	
+	List<ArticleBean> comments = new ArrayList<>();
+	while(rsComment.next()){
+		ArticleBean comment = new ArticleBean();
+	
+		comment.setSeq(rsComment.getInt(1));
+		comment.setParent(rsComment.getInt(2));		
+		comment.setContent(rsComment.getString(6));
+		comment.setUid(rsComment.getString(9));
+		comment.setRegip(rsComment.getString(10));
+		comment.setRdate(rsComment.getString(11));
+		comment.setNick(rsComment.getString(12));
+		
+		comments.add(comment);
+	}
+	
 	// 6단계
+	rsComment.close();
+	psmtComment.close();
 	psmtHit.close();
 	rs.close();	
 	psmt.close();
 	conn.close();
+	
+	// 수정을 대비하기 위한 article객체 세션에 저장
+	session.setAttribute("article", article);
 %>
 <!DOCTYPE html>
 <html lang="en">
@@ -100,29 +126,39 @@
             	
             </script>
             <div>
-                <a href="/Jboard1/proc/delete.jsp?seq=<%= article.getSeq() %>" onclick="return onDelete()" class="btnDelete">삭제</a>
-                <a href="#" class="btnModify">수정</a>
-                <a href="/Jboard1/list.jsp" class="btnList">목록</a>
+            	<% if(mb.getUid().equals(article.getUid())){ %>
+	                <a href="/Jboard1/proc/delete.jsp?seq=<%= article.getSeq() %>" onclick="return onDelete()" class="btnDelete">삭제</a>
+	                <a href="/Jboard1/modify.jsp" class="btnModify">수정</a>
+                <% } %>
+               	 	<a href="/Jboard1/list.jsp" class="btnList">목록</a>
             </div>  
             
             <!-- 댓글리스트 -->
             <section class="commentList">
                 <h3>댓글목록</h3>
-                <% if(article.getComment() > 0){ %>
+                <% for(ArticleBean comment : comments){ %>
                 <article class="comment">
                     <span>
-                        <span>길동이</span>
-                        <span>20-05-13</span>
+                        <span><%= comment.getNick() %></span>
+                        <span><%= comment.getRdate().substring(2, 10) %></span>
                     </span>
-                    <textarea name="comment" readonly>댓글 샘플입니다.</textarea>
+                    <textarea name="comment" readonly><%= comment.getContent() %></textarea>
                     <div>
-                        <a href="#">삭제</a>
+                    <% if(mb.getUid().equals(comment.getUid())){ %>
+                        <a href="/Jboard1/proc/deleteComment.jsp?seq=<%= comment.getSeq() %>&parent=<%= comment.getParent()%>">삭제</a>
                         <a href="#">수정</a>
+                    <% } %>
                     </div>
                 </article>
-                <% } else { %>
-                <p class="empty">등록된 댓글이 없습니다.</p>
-                <% } %>
+                <% 
+                	}
+                
+                	if(comments.size() == 0){
+                %>
+                	<p class="empty">등록된 댓글이 없습니다.</p>
+                <%
+                	}
+                %>
             </section>
 
             <!-- 댓글입력폼 -->
